Add models, normalize namespace
