--- common/src/main/scala/io/kafkamate/common/Models.scala
@@ -1,5 +1,0 @@
-package io.kafkamate.common
-
-object Models {
-
-}
