+++ core/src/main/scala/io/km8/core/utils/Logger.scala
@@ -1,0 +1,17 @@
+package io.km8.core
+package utils
+
+import zio.URLayer
+import zio.clock.Clock
+import zio.console.Console
+import zio.logging._
+
+object Logger {
+
+  lazy val liveLayer: URLayer[Console with Clock, Logging] =
+    Logging.console(
+      logLevel = LogLevel.Info,
+      format = LogFormat.ColoredLogFormat()
+    ) >>> Logging.withRootLoggerName("kafkamate")
+
+}
