+++ core/src/test/scala/io/kafkamate/kafka/consumer/ConsumerSpec.scala
