+++ core/src/test/scala/io/kafkamate/util/HelperSpec.scala
