+++ core/src/test/scala/io/kafkamate/util/KafkaEmbedded.scala
