--- fx/.scalafmt.conf
@@ -1,29 +1,0 @@
-version = "3.1.1"
-runner.dialect = scala3
-
-align.preset = some
-maxColumn = 120
-
-align.tokens = [
-  {code = "%", owner = "Term.ApplyInfix"},
-  {code = "%%", owner = "Term.ApplyInfix"},
-  {code = ":=", owner = "Term.ApplyInfix"},
-  {code = "=>", owner = "Case"}
-]
-
-align.openParenCallSite = false
-
-newlines.source=keep
-newlines.topLevelStatements = [before]
-newlines.beforeMultiline = keep
-newlines.topLevelStatementsMinBreaks = 1
-
-optIn.breakChainOnFirstMethodDot = true
-
-continuationIndent.defnSite = 2
-
-verticalMultiline.atDefnSite = true
-verticalMultiline.arityThreshold = 3
-verticalMultiline.newlineAfterOpenParen = true
-
-danglingParentheses.callSite = false
