--- fx/project/plugins.sbt
@@ -1,3 +1,0 @@
-addSbtPlugin("io.spray" % "sbt-revolver" % "0.9.1")
-
-addSbtPlugin("org.scalameta" % "sbt-scalafmt" % "2.4.3")
