--- fx/src/main/scala/org/github/km8/ui/BaseControl.scala
@@ -1,12 +1,0 @@
-package org.github.km8.ui
-
-import scalafx.scene.Node
-import scalafx.scene.control.Alert.AlertType
-import scalafx.scene.control.{Alert, ButtonType}
-import zio.*
-
-trait BaseControl[R <: Has[_]]:
-  def render: ZIO[R, Throwable, Node]
-
-  def alert(text: Any) =
-    new Alert(AlertType.None, text.toString, ButtonType.OK).show()
