--- fx/src/main/scala/org/github/km8/ui/helpers.scala
@@ -1,9 +1,0 @@
-package org.github.km8.ui
-
-import scalafx.scene.Node
-import scalafx.scene.control.Alert.AlertType
-import scalafx.scene.control.{Alert, ButtonType}
-import zio.*
-
-def alert(text: Any) =
-  new Alert(AlertType.None, text.toString, ButtonType.OK).show()
